% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{all_but_last_split}
\alias{all_but_last_split}
\title{Given a node's address identify the its last split}
\usage{
all_but_last_split(d)
}
\arguments{
\item{d}{A list or character vector of nodes in the tree}
}
\value{
A list
}
\description{
Given a node's address identify the its last split
}
