% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{children}
\alias{children}
\title{Find children of a given set of nodes}
\usage{
children(d, nodes)
}
\arguments{
\item{d}{A list or character vector of nodes in the tree}

\item{nodes}{A list containing all nodes in the tree}
}
\value{
A list, one entry for each element of \code{d}, containing the children
of node \code{d}. For a terminal node the return value is \code{NULL}
}
\description{
Find children of a given set of nodes
}
