% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_method.R
\name{criterion}
\alias{criterion}
\title{HME Selection Criterion}
\usage{
criterion(obj, type = c("aic", "bic", "mse"))
}
\arguments{
\item{obj}{An S3 object of class "hme"}

\item{type}{One of three options: 'aic', 'bic', or 'mse'}
}
\description{
Function to calculate one of several statistics that can be used for model
comparison.
}
