% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{cross_gaussian_hessian}
\alias{cross_gaussian_hessian}
\title{The hessian for parameters of two different expert regressions}
\usage{
cross_gaussian_hessian(
  expert1,
  expert2,
  expert.pars,
  list_priors,
  list_density,
  Y,
  X
)
}
\description{
The hessian for parameters of two different expert regressions
}
