% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{cross_gaussian_multinomial_hessian}
\alias{cross_gaussian_multinomial_hessian}
\title{The hessian for parameters of a gating node and an expert}
\usage{
cross_gaussian_multinomial_hessian(
  node,
  expert,
  expert.pars,
  list_priors,
  list_density,
  Y,
  X,
  Z
)
}
\description{
The hessian for parameters of a gating node and an expert
}
