% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{cross_logistic_hessian}
\alias{cross_logistic_hessian}
\title{The hessian for parameters of two different gating nodes}
\usage{
cross_logistic_hessian(node1, node2, list_priors, list_density, Z)
}
\description{
The hessian for parameters of two different gating nodes
}
