% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{em_algo}
\alias{em_algo}
\title{The EM Algorithm}
\usage{
em_algo(tree, expert_type, posteriors, Y, X, Z, exp.pars, gat.pars)
}
\arguments{
\item{tree}{The tree structure of the HME}

\item{expert_type}{A character identifying the glm type of the experts}

\item{posteriors}{A list of posterior weights from the previous EM step}

\item{Y}{The dependent variable}

\item{X}{The explanatory variables in the expert regressions}

\item{Z}{The explanatory variables in the gating network}

\item{exp.pars}{A list of parameter vectors for each regression}

\item{gat.pars}{A list (of lists) of parameter vectors for each gating node}
}
\description{
Optimize the parameters of the model given the posterior weights constructed
during the E-Step. The E-step is summarized in the posterior weights that are
passed during the function call. The M-stap utilizes the posterior weights
to update the parameters in the model. Using the updated parameter values,
new values for the posterior weights, density estimates, and posterior
weights are returned calculated and returned.
}
