% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_method.R
\name{expert_lik_contr}
\alias{expert_lik_contr}
\title{Likelihood contribution for each expert}
\usage{
expert_lik_contr(experts, densities, weights)
}
\arguments{
\item{experts}{A character vector of expert node names}

\item{densities}{A list containing the density values for all experts in
the tree}

\item{weights}{A list containing all the current split probabilities for
every gating node in the tree}
}
\description{
Likelihood contribution for each expert
}
