% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{gate_path_product}
\alias{gate_path_product}
\title{Find the cumulative product path from an arbitrary gating node to one of its
progeny}
\usage{
gate_path_product(geezer, youngin, gate_prob)
}
\arguments{
\item{geezer}{The address of the gating node closer to the root node}

\item{youngin}{The address of the gating node closer to the expert node}

\item{gate_prob}{The list containing all the current split probabilities for
every gating node in the tree}
}
\description{
Find the cumulative product path from an arbitrary gating node to one of its
progeny
}
