% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{gate_path_values}
\alias{gate_path_values}
\title{Given a list of branches retrieve the gating split probabilities}
\usage{
gate_path_values(branches, gate_prob)
}
\arguments{
\item{branches}{The address of the expert node as a character}

\item{gate_prob}{The list containing all the current split probabilities for
everying gating node in the tree}
}
\value{
A list of matrices with split probabilities
}
\description{
Say gating node 1|0 splits into three directions: 1|0,1 - 1|0,2 - 1|0,3. This
function will retrieve the split probabilities of direction i from node a
}
