% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{gaussian_hessian}
\alias{gaussian_hessian}
\title{The hessian for parameters of the same expert regression}
\usage{
gaussian_hessian(expert, expert.pars, list_priors, list_density, Y, X)
}
\description{
The hessian for parameters of the same expert regression
}
