% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{glm_score}
\alias{glm_score}
\title{Score vector of an expert node}
\usage{
glm_score(expert, expert.pars, family, gate_probs, densities, Y, X)
}
\arguments{
\item{expert}{The name of the expert node}

\item{expert.pars}{The list of parameter vectors for the experts}

\item{family}{The family of glm}

\item{gate_probs}{The list containing all the current prior split probabilities
for every gating node in the tree}

\item{densities}{The list of density values for all experts}

\item{Y}{The dependent variable}

\item{X}{The set of explanatory variables in the expert regressions}
}
\description{
Reference equation 2.13 of the GLM book (pg 41)
}
