% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{init_gate_node_pars}
\alias{init_gate_node_pars}
\title{Supply random initial parameter values for a gating node}
\usage{
init_gate_node_pars(node, tree, n)
}
\arguments{
\item{node}{A gating node name}

\item{tree}{A list of the entire tree structure}

\item{n}{The number of parameters to provide initial values}
}
\description{
Supply random initial parameter values for a gating node
}
