% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{inter_node_paths}
\alias{inter_node_paths}
\title{Find the cascading split probabilities from an arbitrary gating node to one
of its progeny}
\usage{
inter_node_paths(geezer, youngin, gate_prob)
}
\arguments{
\item{geezer}{The address of the gating node closer to the root node}

\item{youngin}{The address of the gating node closer to the expert node}

\item{gate_prob}{The list containing all the current split probabilities for
every gating node in the tree}
}
\description{
Find the cascading split probabilities from an arbitrary gating node to one
of its progeny
}
