% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{irls_logit}
\alias{irls_logit}
\title{Fitting procedure for a binomial logistic regression}
\usage{
irls_logit(X, H, w = NULL, maxit = 25, tol = 1e-08)
}
\arguments{
\item{X}{The matrix of explanatory variables}

\item{H}{The target variable for the logistic regression}

\item{w}{A numeric vector of weights attached to each observation in X}

\item{maxit}{The maximum number of iterations for the IRLS algorithm}

\item{tol}{Numeric indicating the convergence tolerance}
}
\description{
Fitting procedure for a binomial logistic regression
}
