% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{is_terminal}
\alias{is_terminal}
\title{Test if a node is an expert}
\usage{
is_terminal(d, nodes)
}
\arguments{
\item{d}{A list or character vector of nodes in the tree}

\item{nodes}{A list containing all nodes in the tree}
}
\value{
A list of Boolean values, one entry for each element of \code{d},
testing if each element of \code{d} is an expert node
}
\description{
Test if a node is an expert
}
