% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{joint_posterior_weight}
\alias{joint_posterior_weight}
\title{Cumulative product of posterior weights from the root to an arbitrary node}
\usage{
joint_posterior_weight(node, post_weights)
}
\arguments{
\item{node}{An arbitrary node. Can be either a gating node or an expert}

\item{post_weights}{The list containing all the posterior split probabilities
for every gating node in the tree}
}
\description{
Cumulative product of posterior weights from the root to an arbitrary node
}
