% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_method.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Likelihood contributions}
\usage{
log_likelihood(treestr, gate_prob, densities)
}
\arguments{
\item{treestr}{A list of the entire tree structure}

\item{gate_prob}{The list containing all the current prior split probabilities
for every gating node in the tree}

\item{densities}{The list containing the density values for all experts in
the tree}
}
\value{
The log-likelihood of the model
}
\description{
Calculates the matrix of likelihood contributions for all experts in the tree.
Used internally
}
