% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{logistic_score}
\alias{logistic_score}
\title{Score vector of a gating node}
\usage{
logistic_score(node, gate_probs, densities, Z)
}
\arguments{
\item{node}{The name of the gating node}

\item{gate_probs}{The list containing all the current prior split probabilities
for every gating node in the tree}

\item{densities}{The list of density values for all experts}

\item{Z}{The set of explanatory variables in the gating node}
}
\description{
Reference equation 1.27 of the dissertation (pg 18)
}
