% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_method.R
\name{logLik.hme}
\alias{logLik.hme}
\title{Likelihood value for the HME}
\usage{
\method{logLik}{hme}(obj)
}
\arguments{
\item{obj}{An S3 object of class "hme"}
}
\description{
S3 method to calculate the likelihood of a fitted HME model object. The
value comes from the fitted parameters of the model and not the vector of
historical likelihood values saved from model optimization
}
