% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{napply}
\alias{napply}
\title{Wrapper for lapply where x is a list of characters. Ensures the values of \code{x}
are passed as names to the returned list}
\usage{
napply(x, f_, ...)
}
\arguments{
\item{x}{A list}

\item{f_}{The function to apply}

\item{...}{Additional variables to pass to `lapply`}
}
\value{
A list
}
\description{
Wrapper for lapply where x is a list of characters. Ensures the values of \code{x}
are passed as names to the returned list
}
