% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{Omega_0}
\alias{Omega_0}
\title{Rate of change of likelihood to gating parameters}
\usage{
Omega_0(node, gate_probs, GD)
}
\arguments{
\item{node}{The name of the gating node}

\item{gate_probs}{The list containing all the current prior split probabilities
for every gating node in the tree}

\item{GD}{The list of likelihood contributions for all experts}
}
\description{
Reference equation 1.28 of the dissertation (pg 18)
}
