% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{optimize_block}
\alias{optimize_block}
\title{Optimize a node in the HME}
\usage{
optimize_block(Y, X, wts, node_type = c("binomial", "multinomial", "gaussian"))
}
\arguments{
\item{Y}{The dependent variable}

\item{X}{The set of explanatory variables}

\item{node_type}{The type of node that is being optimized}
}
\value{
A numeric vector of updated parameter values
}
\description{
This function will optimize each node in the HME. Iterative re-weighted least
squares (IRLS) is used as the optimization technique. A hand-written implementation
is used for the gating node while glm.fit is used for the expert regressions
}
