% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{par_to_gate_paths}
\alias{par_to_gate_paths}
\title{Calculate the multinomial split probabilities of an arbitrary gating node
given the data and the gating parameters}
\usage{
par_to_gate_paths(node, gate_par_list, X)
}
\arguments{
\item{node}{The address of the gating node}

\item{gate_par_list}{A list (of lists) containing the parameters of the gating nodes}

\item{X}{The design matrix of the gating network}
}
\value{
A matrix of multinomial probabilities for node \code{d}
}
\description{
Calculate the multinomial split probabilities of an arbitrary gating node
given the data and the gating parameters
}
