% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{parent}
\alias{parent}
\title{Find the parent of a given set of nodes}
\usage{
parent(d, nodes)
}
\arguments{
\item{d}{A list or character vector of nodes in the tree}

\item{nodes}{A list containing all nodes in the tree}
}
\value{
A list, one entry for each element of \code{d}, containing the parent
of node \code{d}. For the root node the return value is \code{NULL}
}
\description{
Find the parent of a given set of nodes
}
