% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{prior_weights}
\alias{prior_weights}
\title{Calculate the prior weight for a gating node}
\usage{
prior_weights(node, gate_prob)
}
\arguments{
\item{node}{An arbitrary node in the tree. Can be a gating node or an expert
node}

\item{gate_prob}{The list containing all the current split probabilities for
every gating node in the tree}
}
\description{
Calculate the prior weight for a gating node
}
