% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{progeny}
\alias{progeny}
\title{Find the progeny of a given set of nodes}
\usage{
progeny(d, nodes)
}
\arguments{
\item{d}{A list or character vector of nodes in the tree}

\item{nodes}{A list containing all nodes in the tree}
}
\value{
A list, one entry for each element of \code{d}, providing the ancestry
of each node in \code{d} back to the root node. For the root node the return
value is \code{NULL}
}
\description{
Find the progeny of a given set of nodes
}
