% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{root_to_node_weight}
\alias{root_to_node_weight}
\title{Calculate the cumulative weight}
\usage{
root_to_node_weight(node, gate_prob)
}
\arguments{
\item{node}{An arbitrary node in the tree. Can be a gating node or an expert
node}

\item{gate_prob}{The list containing all the current split probabilities for
every gating node in the tree}
}
\description{
Calculate the cumulative product from the root node to an arbitrary node
}
