% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{sandwich_vcov}
\alias{sandwich_vcov}
\title{Sandwich Var-Cov Matrix}
\usage{
sandwich_vcov(
  gte.nms,
  exp.nms,
  scores,
  densities,
  weights,
  gate.pars,
  exp.pars,
  Y,
  X,
  Z,
  N
)
}
\description{
Sandwich Var-Cov Matrix
}
