% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\name{summary.hme}
\alias{summary.hme}
\title{Summary Method}
\usage{
\method{summary}{hme}(
  obj,
  vcv = c("OPG", "hessian", "sandwich"),
  type = c("all", "experts", "gates")
)
}
\arguments{
\item{obj}{An S3 object of class "hme"}

\item{vcv}{The type of variance-covariance matrix to use for standard errors.
Defaults to the outer product of the gradient (OPG)}

\item{type}{Select which nodes of the HME to provide summary reports}
}
\description{
Summarize the fitted parameters of the HME
}
