% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voung_selection.R
\name{voung_selection}
\alias{voung_selection}
\title{Voung's Likelihood Ratio Test}
\usage{
voung_selection(
  hme1,
  hme2,
  var_test_alpha = 0.05,
  model_test_alpha = 0.05,
  adjustedLR = TRUE
)
}
\arguments{
\item{hme1}{An S3 object of class 'hme'}

\item{hme2}{An S3 object of class 'hme'}

\item{var_test_alpha}{Significance level for the LR test}

\item{adjustedLR}{Boolean value to use an adjusted LR test}
}
\description{
Model comparison using Voung's likelihood ratio test. See his 1989 publication
in Econometrica. In order to use this comparison test, the Hessian for both
HME models needs to be calculated
}
