% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{expert_index}
\alias{expert_index}
\title{Given a list of nodes return identify which ones are expert nodes}
\usage{
expert_index(d)
}
\arguments{
\item{d}{A list or character vector of all nodes in the tree}
}
\value{
Given a node and its address return the direction of the last split
}
\description{
Given a list of nodes return identify which ones are expert nodes
}
