% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{expert_lik_contr}
\alias{expert_lik_contr}
\title{Calculate the likelihood contribution for a set of experts}
\usage{
expert_lik_contr(experts, densities, gate_prob)
}
\arguments{
\item{experts}{A character vector of expert node names}

\item{densities}{he list containing the density values for all experts in
the tree}

\item{gate_prob}{The list containing all the current split probabilities for
every gating node in the tree}
}
\description{
Calculate the likelihood contribution for a set of experts
}
